+++ src/main/java/com/sixmac/controller/api/StadiumApi.java
@@ -72,7 +72,7 @@ public class StadiumApi extends CommonController {
      * @apiParam {Integer} type 球赛类型 N人制 N代表数量
      * @apiParam {Double} longitude 经度  <必传/>
      * @apiParam {Double} latitude 纬度  <必传/>
-     * @apiParam {Integer} areaId 区域ID
+     * @apiParam {Long} areaId 区域ID
      * @apiParam {Integer} pageNum 当前页
      * @apiParam {Integer} pageSize 每页显示数
      * @apiSuccess {Object}  list 球场列表
@@ -94,7 +94,7 @@ public class StadiumApi extends CommonController {
     public void list(HttpServletResponse response,
                      Double longitude,
                      Double latitude,
-                     Integer areaId,
+                     Long areaId,
                      Integer type,
                      Integer pageNum,
                      Integer pageSize) throws IOException {
@@ -125,23 +125,23 @@ public class StadiumApi extends CommonController {
      * @api {post} /api/stadium/info 球场详情
      * @apiName stadium.info
      * @apiGroup stadium
-     * @apiParam {Integer} stadiumId 球场ID  <必传/>
+     * @apiParam {Long} stadiumId 球场ID  <必传/>
      * @apiSuccess {Object}  stadium 球场
-     * @apiSuccess {Integer} stadium.id 球场id
+     * @apiSuccess {Long} stadium.id 球场id
      * @apiSuccess {String} stadium.name 球场名称
      * @apiSuccess {String} stadium.areaName 球场地区名字
      * @apiSuccess {String} stadium.address 球场地址
      * @apiSuccess {String} stadium.avater 球场封面
      * @apiSuccess {String} stadium.siteType 场地类型
      * @apiSuccess {String} stadium.sodType 草皮类型
-     * @apiSuccess {Integer} stadium.light 灯光(0：有灯光 1：没有灯光)
+     * @apiSuccess {String} stadium.light 灯光
      * @apiSuccess {Integer} stadium.park 停车场（0：有停车场 1；没有停车场）
      * @apiSuccess {String} stadium.giving 赠送
      * @apiSuccess {String} stadium.description 球场简介
      * @apiSuccess {Object} sites 球场场地
      */
     @RequestMapping(value = "/info")
-    public void info(HttpServletResponse response, Integer stadiumId) throws ParseException {
+    public void info(HttpServletResponse response, Long stadiumId) throws ParseException {
 
         if (null == stadiumId ) {
             WebUtil.printJson(response, new Result(false).msg(ErrorCode.ERROR_CODE_0002));
@@ -189,15 +189,15 @@ public class StadiumApi extends CommonController {
      * @api {post} /api/stadium/order 公共球场约散客
      * @apiName stadium.order
      * @apiGroup stadium
-     * @apiParam {Integer} stadiumId 球场ID <必传/>
+     * @apiParam {Long} stadiumId 球场ID <必传/>
      * @apiSuccess {Object}  stadium 球场
-     * @apiSuccess {Integer} stadium.id 球场id
+     * @apiSuccess {Long} stadium.id 球场id
      * @apiSuccess {String} stadium.name 球场名称
      * @apiSuccess {String} stadium.areaName 球场地区名字
      * @apiSuccess {String} stadium.avater 球场封面
      */
     @RequestMapping(value = "/order")
-    public void order(HttpServletResponse response, Integer stadiumId) {
+    public void order(HttpServletResponse response, Long stadiumId) {
 
         if (null == stadiumId ) {
             WebUtil.printJson(response, new Result(false).msg(ErrorCode.ERROR_CODE_0002));
@@ -218,21 +218,21 @@ public class StadiumApi extends CommonController {
      * @api {post} /api/stadium/publish 公共球场发布
      * @apiName stadium.publish
      * @apiGroup stadium
-     * @apiParam {Integer} stadiumId 球场ID <必传/>
-     * @apiParam {Integer} userId 用户ID <必传/>
+     * @apiParam {Long} stadiumId 球场ID <必传/>
+     * @apiParam {Long} userId 用户ID <必传/>
      * @apiParam {String} title 标题  <必传/>
      * @apiParam {Long} time 约球时间 <必传/>
      *
      * @apiSuccess {Object}  stadium 球场
-     * @apiSuccess {Integer} stadium.id 球场id
+     * @apiSuccess {Long} stadium.id 球场id
      * @apiSuccess {String} stadium.name 球场名称
-     * @apiSuccess {Integer} stadium.cityId 球场地区
+     * @apiSuccess {Long} stadium.cityId 球场地区
      * @apiSuccess {String} stadium.avater 球场封面
      */
     @RequestMapping(value = "/publish")
     public void publish(HttpServletResponse response,
-                        Integer userId,
-                        Integer stadiumId,
+                        Long userId,
+                        Long stadiumId,
                         String title,
                         Long time) {
 
@@ -265,7 +265,7 @@ public class StadiumApi extends CommonController {
      * @api {post} /api/stadium/siteSelect 场次选择
      * @apiName stadium.siteSelect
      * @apiGroup stadium
-     * @apiParam {Integer} stadiumId 球场ID <必传/>
+     * @apiParam {Long} stadiumId 球场ID <必传/>
      * @apiParam {Long} time 当天时间戳 <必传/>
      *
      * @apiSuccess {Object}  site 场地
@@ -274,7 +274,7 @@ public class StadiumApi extends CommonController {
      * @apiSuccess {String} list 预定字符串 (0：不可预定 1：可预订)
      */
     @RequestMapping(value = "/siteSelect")
-    public void siteSelect(HttpServletResponse response, Long time, Integer stadiumId) {
+    public void siteSelect(HttpServletResponse response, Long time, Long stadiumId) {
 
         if (null == stadiumId || time == null ) {
             WebUtil.printJson(response, new Result(false).msg(ErrorCode.ERROR_CODE_0002));
@@ -316,14 +316,14 @@ public class StadiumApi extends CommonController {
      * @api {post} /api/stadium/siteOrder 场地预订
      * @apiName stadium.siteOrder
      * @apiGroup stadium
-     * @apiParam {Integer} siteId 场地ID <必传/>
+     * @apiParam {Long} siteId 场地ID <必传/>
      * @apiParam {Integer} status 运动保险状态（0：不买  1：买） <必传/>
      * @apiParam {Long} time 时间戳  <必传/>
      * @apiParam {Integer} start 开始时间点  <必传/>
      * @apiParam {Integer} end 结束时间点  <必传/>
      *
      * @apiSuccess {Object}  siteTime 场地预定
-     * @apiSuccess {Integer} siteTime.id 预定id
+     * @apiSuccess {Long} siteTime.id 预定id
      * @apiSuccess {Object} siteTime.site 预定场地
      * @apiSuccess {String} siteTime.site.code 场地编号
      * @apiSuccess {String} siteTime.site.type 场地类型
@@ -336,7 +336,7 @@ public class StadiumApi extends CommonController {
      */
     @RequestMapping(value = "/siteOrder")
     public void siteOrder(HttpServletResponse response,
-                          Integer siteId,
+                          Long siteId,
                           Integer status,
                           Long time,
                           Integer start,
@@ -378,14 +378,14 @@ public class StadiumApi extends CommonController {
      * @api {post} /api/stadium/pay 散客支付订单
      * @apiName stadium.pay
      * @apiGroup stadium
-     * @apiParam {Integer} siteTimeId 预定场地ID <必传/>
-     * @apiParam {Integer} userId 用户ID <必传/>
-     * @apiParam {Integer} insuranceId 保险ID
+     * @apiParam {Long} siteTimeId 预定场地ID <必传/>
+     * @apiParam {Long} userId 用户ID <必传/>
+     * @apiParam {Long} insuranceId 保险ID
      * @apiParam {Integer} num 购买保险数
      *
      * @apiSuccess {Object}  site 场地
      * @apiSuccess {Object}  sysInsurance 保险
-     * @apiSuccess {Integer} sysInsurance.id 保险id
+     * @apiSuccess {Long} sysInsurance.id 保险id
      * @apiSuccess {Integer} sysInsurance.name 保险名称
      * @apiSuccess {String} sysInsurance.price 保险金额
      * @apiSuccess {Integer} siteMoney 场地费
@@ -393,13 +393,13 @@ public class StadiumApi extends CommonController {
      * @apiSuccess {Integer} preferente 会员折扣
      * @apiSuccess {String} money 总金额
      * @apiSuccess {Object}  reserve 散客预定（约球）
-     * @apiSuccess {Integer} reserve.id 预定id
+     * @apiSuccess {Long} reserve.id 预定id
      */
     @RequestMapping(value = "/pay")
     public void pay(HttpServletResponse response,
-                    Integer userId,
-                    Integer siteTimeId,
-                    Integer insuranceId,
+                    Long userId,
+                    Long siteTimeId,
+                    Long insuranceId,
                     Integer num,
                     Double preferente,
                     Double money) {
@@ -482,7 +482,7 @@ public class StadiumApi extends CommonController {
      * @api {post} /api/stadium/payConfirm 散客确认支付订单
      * @apiName stadium.payConfirm
      * @apiGroup stadium
-     * @apiParam {Integer} reserveId 预定ID <必传/>
+     * @apiParam {Long} reserveId 预定ID <必传/>
      * @apiParam {Integer} type 支付类型（0：全额支付  1：AA支付） <必传/>
      * @apiSuccess {Object} order 订单
      * @apiSuccess {String} order.userName 用户昵称
@@ -491,7 +491,7 @@ public class StadiumApi extends CommonController {
      * @apiSuccess {Long} order.sn 订单号
      */
     @RequestMapping(value = "/payConfirm")
-    public void payConfirm(HttpServletResponse response, Integer reserveId, Integer type, Double money) {
+    public void payConfirm(HttpServletResponse response, Long reserveId, Integer type, Double money) {
 
         if (null == reserveId || type == null ) {
             WebUtil.printJson(response, new Result(false).msg(ErrorCode.ERROR_CODE_0002));
@@ -530,26 +530,26 @@ public class StadiumApi extends CommonController {
      * @api {post} /api/stadium/teamPay 球队支付订单
      * @apiName stadium.teamPay
      * @apiGroup stadium
-     * @apiParam {Integer} siteTimeId 预定场地ID <必传/>
-     * @apiParam {Integer} userId 用户ID <必传/>
-     * @apiParam {Integer} insuranceId 保险ID
+     * @apiParam {Long} siteTimeId 预定场地ID <必传/>
+     * @apiParam {Long} userId 用户ID <必传/>
+     * @apiParam {Long} insuranceId 保险ID
      * @apiParam {Integer} num 购买保险数
      * @apiSuccess {Object}  site 场地
      * @apiSuccess {Object}  sysInsurance 保险
-     * @apiSuccess {Integer} sysInsurance.name 保险名称
-     * @apiSuccess {String} sysInsurance.price 保险金额
-     * @apiSuccess {Integer} siteMoney 场地费
-     * @apiSuccess {Object}  vipNum 会员等级
-     * @apiSuccess {Integer} preferente 会员折扣
-     * @apiSuccess {String} money 总金额
+     * @apiSuccess {String} sysInsurance.name 保险名称
+     * @apiSuccess {Double} sysInsurance.price 保险金额
+     * @apiSuccess {Double} siteMoney 场地费
+     * @apiSuccess {Integer}  vipNum 会员等级
+     * @apiSuccess {Double} preferente 会员折扣
+     * @apiSuccess {Double} money 总金额
      * @apiSuccess {Object}  reserveTeam 球队预定
-     * @apiSuccess {Integer} reserveTeam.id 预定id
+     * @apiSuccess {Long} reserveTeam.id 预定id
      */
     @RequestMapping(value = "/teamPay")
     public void teamPay(HttpServletResponse response,
-                        Integer userId,
-                        Integer siteTimeId,
-                        Integer insuranceId,
+                        Long userId,
+                        Long siteTimeId,
+                        Long insuranceId,
                         Integer num,
                         Double preferente,
                         Double money) {
@@ -620,7 +620,7 @@ public class StadiumApi extends CommonController {
      * @api {post} /api/stadium/teamPayConfirm 球队确认支付订单
      * @apiName stadium.teamPayConfirm
      * @apiGroup stadium
-     * @apiParam {Integer} reserveTeamId 预定ID <必传/>
+     * @apiParam {Long} reserveTeamId 预定ID <必传/>
      * @apiSuccess {Object} order 订单
      * @apiSuccess {String} order.userName 用户昵称
      * @apiSuccess {String} order.stadiumName 球场名称
@@ -628,7 +628,7 @@ public class StadiumApi extends CommonController {
      * @apiSuccess {Long} order.sn 订单号
      */
     @RequestMapping(value = "/teamPayConfirm")
-    public void teamPayConfirm(HttpServletResponse response, Integer reserveTeamId, Double money) {
+    public void teamPayConfirm(HttpServletResponse response, Long reserveTeamId, Double money) {
 
         if (null == reserveTeamId ) {
             WebUtil.printJson(response, new Result(false).msg(ErrorCode.ERROR_CODE_0002));
