+++ src/main/java/com/sixmac/dao/ActivityDao.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 /**
  * Created by Administrator on 2016/5/19 0019 上午 10:08.
  */
-public interface ActivityDao extends JpaRepository<Activity, Integer>, JpaSpecificationExecutor<Activity> {
+public interface ActivityDao extends JpaRepository<Activity, Long>, JpaSpecificationExecutor<Activity> {
 
 }
